added react router made layout of home page
