+++ .gitignore
@@ -1,3 +1,4 @@
 .gitignore
 .package-lock.json
 .package.json
+node_modules
