+++ frontend/src/components/Home/Home.jsx
@@ -1,0 +1,121 @@
+import React from "react";
+import { CiSearch } from "react-icons/ci";
+import { CiHome } from "react-icons/ci";
+import { MdOutlineLibraryMusic } from "react-icons/md";
+import { IoMdPhotos } from "react-icons/io";
+import { HiMiniMusicalNote } from "react-icons/hi2";
+import { TbPlaylist } from "react-icons/tb";
+import { RiDeleteBin6Line } from "react-icons/ri";
+import { FaChevronLeft } from "react-icons/fa";
+import { FaChevronRight } from "react-icons/fa";
+import { MdAccountCircle } from "react-icons/md";
+import { IoIosNotifications } from "react-icons/io";
+
+const Home = () => {
+    return (
+        <div className="w-full h-full flex items-center justify-center bg-gradient-to-20 mx-auto">
+
+            <div className="sidebar w-1/5 h-full">
+                <div className="search w-full h-1/6 flex justify-center items-center ">
+                    <div className="w-4/5 h-1/5 bg-white bg-opacity-35 rounded-full border-none flex justify-center items-center p-2">
+                        <CiSearch size={30} color={"white"} />
+                        <input type="text" placeholder="Search" className=" search-bar w-full h-full  bg-transparent p-2 font-bold" />
+                    </div>
+
+                </div>
+
+                <div className="navs w-full h-4/5">
+                    <div className="menu-library w-full h-1/2 flex flex-col items-left justify-around   text-white text-xl">
+                        <h1 className="px-8 font-bold">MENU</h1>
+                        <div className="flex justify-left items-center px-8"> <CiHome /> <span className="p-2"></span>Home</div>
+                        <h1 className="px-8 font-bold">LIBRARY</h1>
+                        <div className="flex justify-left items-center px-8"> <IoMdPhotos /> <span className="p-2"></span>Albums</div>
+                        <div className="flex justify-left items-center px-8"> <MdOutlineLibraryMusic /> <span className="p-2"></span>Songs</div>
+                        <div className="flex justify-left items-center px-8"> <HiMiniMusicalNote /> <span className="p-2"></span>Artists</div>
+
+
+                    </div>
+
+                    <div className="playlist w-full h-2/5 text-white text-xl flex flex-col justify-around items-left">
+                        <h1 className="px-8 font-bold">PLAYLISTS</h1>
+                        <div className="px-8 flex justify-left items-center"><TbPlaylist /> <span className="px-2">playlist 1</span> <RiDeleteBin6Line /></div>
+                        <div className="px-8 flex justify-left items-center"><TbPlaylist /> <span className="px-2">playlist 2</span> <RiDeleteBin6Line /></div>
+                        <div className="px-8 flex justify-left items-center"><TbPlaylist /> <span className="px-2">playlist 3</span> <RiDeleteBin6Line /></div>
+
+                    </div>
+
+                </div>
+
+            </div>
+
+            <div className="main w-3/5 h-full flex-col justify-around items-left">
+
+                <div className="top-left-icons flex justify-left items-center p-4">
+                    <div className="left-icon bg-white bg-opacity-35 rounded-full flex justify-center items-center p-2 ">
+                        <FaChevronLeft size={30} color={"white"} />
+                    </div> <span className="p-2"></span>
+                    <div className="right-icon bg-white bg-opacity-35 rounded-full flex justify-center items-center p-2">
+                        <FaChevronRight size={30} color={"white"} />
+                    </div>
+                </div>
+
+                <div className="w-full h-2/5">
+
+                    <div className="head flex justify-stretch items-center w-full h-auto">
+                        <h1 className="text-white text-3xl font-bold w-1/2 h-auto">Recently Played</h1>
+                        <h1 className="text-white text-2xl font-bold w-1/2 h-auto text-right">See All</h1>
+                    </div>
+
+
+                    <div className="recents-card w-full h-4/5 bg-white bg-opacity-35 rounded-2xl flex justify-center items-center p-2 mt-2">
+
+                    </div>
+
+                </div>
+
+                <div className="w-full h-2/5 mt-4">
+
+                    <div className="head flex justify-stretch items-center w-full h-auto">
+                        <div className="w-1/2 h-auto"><h1 className="text-white text-xl font-bold w-full h-auto">Global</h1>
+                            <h1 className="text-white text-2xl font-bold w-full h-auto ">Top 10</h1></div>
+                        <h1 className="text-white text-2xl font-bold w-1/2 h-auto text-right">See All</h1>
+                    </div>
+
+
+                    <div className="recents-card w-full h-4/5 bg-white bg-opacity-35 rounded-2xl flex justify-center items-center p-2 mt-2">
+
+                    </div>
+
+                </div>
+
+
+
+
+
+
+            </div>
+
+        <div className="home-player w-2/5 h-full flex flex-col justify-center items-center">
+            <div className="top-right-icons w-full h-1/5 flex justify-end items-start p-4 ">
+                    <div className="left-icon bg-white bg-opacity-35 rounded-full flex justify-center items-center p-2 ">
+                    <MdAccountCircle  size = {40} color = {"white"}/>
+                    </div> <span className="p-2"></span>
+                    <div className="right-icon bg-white bg-opacity-35 rounded-full flex justify-center items-center p-2 ">
+                    <IoIosNotifications size = {40} color = {"white"}/>
+                    </div>
+            </div>
+
+            <div className = "h-4/5 w-full flex flex-col justify-evenly items-center">
+
+            <div className="player w-3/4 h-2/5 bg-white bg-opacity-35 rounded-2xl flex flex-col justify-center items-center p-2"> 
+            <h1 className = "text-lg font-bold text-white w-full text-left p-2">Top Artists</h1>
+            <div className="list w-full h-4/5"></div>
+            </div>
+            <div className="player w-3/4 h-1/2 bg-white bg-opacity-35 rounded-2xl flex justify-center items-center p-2 mt-2"> </div>
+            </div>
+        </div>
+    </div>
+    );
+}
+
+export default Home;
